+++ horizons/messaging/messagebus.py
@@ -20,7 +20,7 @@
 # ###################################################
 
 import logging
-from collections import defaultdict, deque
+from collections import defaultdict
 
 from horizons.util.python.singleton import Singleton
 
@@ -91,82 +91,3 @@ class MessageBus(object):
 
 		# suicide, next instance will be created on demand
 		self.__class__.destroy_instance()
-
-
-class QueuingMessageBus(MessageBus):
-	"""The QueuingMessageBus class is used to send Message instances from a sender to
-	one or multiple recipients, with the additional property that messages will
-	be saved to a queue if no callback is subscribed at the time they are sent."""
-	
-	def __init__(self):
-		MessageBus.__init__(self)
-		# Queue up messages if there is no registered subscriber
-		self.message_queue = defaultdict(deque)
-	
-	def subscribe_globally(self, messagetype, callback):
-		MessageBus.subscribe_globally(self, messagetype, callback)
-		
-		while len(self.message_queue[messagetype]):
-			self.broadcast(self.message_queue[messagetype].popleft())
-
-	def subscribe_locally(self, messagetype, instance, callback):
-		MessageBus.subscribe_locally(self, messagetype, instance, callback)
-		
-		for message in self.message_queue[messagetype]:
-			if (message, message.sender) == (messagetype, instance):
-				self.broadcast(message)
-				self.message_queue[messagetype].remove(message)
-
-	def broadcast(self, message):
-		messagetype = message.__class__
-		pair = (messagetype, message.sender)
-		
-		# check if the message will go anywhere, if not, then queue it
-		if not len(self.global_receivers[messagetype]) and not len(self.local_receivers[pair]):
-			self.message_queue[messagetype].append(message)
-		else:
-			MessageBus.broadcast(self, message)
-
-	def clear(self, messagetype):
-		self.message_queue[messagetype].clear()
-
-	def queue_len(self, messagetype):
-		return len(self.message_queue[messagetype])
-
-
-class SimpleMessageBus(object):
-	"""Manages registration and calling of callbacks when events (strings) occur.
-
-	Example:
-
-		bus = SimpleMessageBus(('foo', 'bar'))
-		bus.subscribe('foo', cb)
-
-		bus.broadcast('foo')  # cb will be called
-	"""
-
-	def __init__(self, message_types):
-		self._message_types = message_types
-		self._callbacks = defaultdict(list)
-
-	def subscribe(self, type, callback):
-		if type not in self._message_types:
-			raise TypeError("Unsupported type")
-		if callback in self._callbacks[type]:
-			raise Exception("Callback %s already subscribed to %s" % (callback, type))
-
-		self._callbacks[type].append(callback)
-
-	def unsubscribe(self, type, callback):
-		self._callbacks[type].remove(callback)
-
-	def discard(self, type, callback):
-		if callback in self._callbacks[type]:
-			self._callbacks[type].remove(callback)
-
-	def broadcast(self, type, *args, **kwargs):
-		if type not in self._message_types:
-			return
-
-		for cb in self._callbacks[type]:
-			cb(*args, **kwargs)
